[Refactor/#92] ApiError 확장함수 core로 이동
