+++ app/src/main/java/com/swyp/firsttodo/presentation/auth/LoginViewModel.kt
@@ -6,12 +6,12 @@ import androidx.navigation.toRoute
 import com.swyp.firsttodo.BuildConfig
 import com.swyp.firsttodo.core.base.Async
 import com.swyp.firsttodo.core.base.BaseViewModel
+import com.swyp.firsttodo.core.common.extension.snackbarMsg
 import com.swyp.firsttodo.core.network.model.ApiError
 import com.swyp.firsttodo.domain.model.SocialType
 import com.swyp.firsttodo.domain.usecase.auth.SocialLoginUseCase
 import com.swyp.firsttodo.presentation.auth.launcher.GoogleLoginResult
 import com.swyp.firsttodo.presentation.auth.navigation.AuthRoute
-import com.swyp.firsttodo.presentation.common.extension.snackbarMsg
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
 import timber.log.Timber
