+++ app/src/main/java/com/swyp/firsttodo/presentation/hamburger/family/FamilyViewModel.kt
@@ -3,9 +3,9 @@ package com.swyp.firsttodo.presentation.hamburger.family
 import androidx.lifecycle.viewModelScope
 import com.swyp.firsttodo.core.base.Async
 import com.swyp.firsttodo.core.base.BaseViewModel
+import com.swyp.firsttodo.core.common.extension.snackbarMsg
 import com.swyp.firsttodo.core.network.model.ApiError
 import com.swyp.firsttodo.domain.repository.FamilyRepository
-import com.swyp.firsttodo.presentation.common.extension.snackbarMsg
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
 import javax.inject.Inject
